refactor(agent): move RunOrchestrator and domain types to soliplex_client
Move 5 source files and 5 test files from soliplex_agent to
soliplex_client. This places the run state machine and supporting
domain types (ThreadKey, FailureReason) in the lower-level client
package where they belong.

A new public entry point (package:soliplex_client/run.dart) exports
the moved types without adding them to the main barrel, avoiding
name collisions with the app's own ThreadKey and run-state types.
soliplex_agent re-exports via its barrel for backward compatibility.

Also removes PlatformConstraints from RunOrchestrator constructor
(platform guards remain in AgentRuntime where they belong).

Moved: run_orchestrator, run_state, error_classifier -> application/
Moved: failure_reason, thread_key -> domain/
