+++ packages/soliplex_agent/lib/src/runtime/agent_runtime.dart
@@ -2,11 +2,10 @@ import 'dart:async';
 
 import 'package:soliplex_agent/src/host/platform_constraints.dart';
 import 'package:soliplex_agent/src/models/agent_result.dart';
-import 'package:soliplex_agent/src/models/thread_key.dart';
-import 'package:soliplex_agent/src/run/run_orchestrator.dart';
 import 'package:soliplex_agent/src/runtime/agent_session.dart';
 import 'package:soliplex_agent/src/runtime/agent_session_state.dart';
 import 'package:soliplex_agent/src/tools/tool_registry_resolver.dart';
+import 'package:soliplex_client/run.dart';
 import 'package:soliplex_client/soliplex_client.dart';
 import 'package:soliplex_logging/soliplex_logging.dart';
 
@@ -215,7 +214,6 @@ class AgentRuntime {
       api: _api,
       agUiClient: _agUiClient,
       toolRegistry: toolRegistry,
-      platformConstraints: _platform,
       logger: _logger,
     );
     return AgentSession(
