+++ packages/soliplex_agent/lib/src/runtime/agent_session.dart
@@ -2,11 +2,8 @@ import 'dart:async';
 
 import 'package:meta/meta.dart';
 import 'package:soliplex_agent/src/models/agent_result.dart';
-import 'package:soliplex_agent/src/models/failure_reason.dart';
-import 'package:soliplex_agent/src/models/thread_key.dart';
-import 'package:soliplex_agent/src/run/run_orchestrator.dart';
-import 'package:soliplex_agent/src/run/run_state.dart';
 import 'package:soliplex_agent/src/runtime/agent_session_state.dart';
+import 'package:soliplex_client/run.dart';
 import 'package:soliplex_client/soliplex_client.dart';
 import 'package:soliplex_logging/soliplex_logging.dart';
