+++ packages/soliplex_agent/test/runtime/agent_session_test.dart
@@ -91,7 +91,6 @@ AgentSession createSession({
     api: api,
     agUiClient: agUiClient,
     toolRegistry: registry,
-    platformConstraints: const NativePlatformConstraints(),
     logger: logger,
   );
   return AgentSession(
