+++ packages/soliplex_client/lib/src/domain/failure_reason.dart
