+++ packages/soliplex_client/lib/src/domain/thread_key.dart
