+++ packages/soliplex_client/test/application/error_classifier_test.dart
@@ -1,5 +1,6 @@
-import 'package:soliplex_agent/soliplex_agent.dart';
 import 'package:soliplex_client/soliplex_client.dart';
+import 'package:soliplex_client/src/application/error_classifier.dart';
+import 'package:soliplex_client/src/domain/failure_reason.dart';
 import 'package:test/test.dart';
 
 void main() {
