+++ packages/soliplex_client/test/application/run_orchestrator_test.dart
@@ -1,8 +1,11 @@
 import 'dart:async';
 
 import 'package:mocktail/mocktail.dart';
-import 'package:soliplex_agent/soliplex_agent.dart';
 import 'package:soliplex_client/soliplex_client.dart';
+import 'package:soliplex_client/src/application/run_orchestrator.dart';
+import 'package:soliplex_client/src/application/run_state.dart';
+import 'package:soliplex_client/src/domain/failure_reason.dart';
+import 'package:soliplex_client/src/domain/thread_key.dart';
 import 'package:soliplex_logging/soliplex_logging.dart';
 import 'package:test/test.dart';
 
@@ -99,7 +102,6 @@ void main() {
       api: api,
       agUiClient: agUiClient,
       toolRegistry: const ToolRegistry(),
-      platformConstraints: const NativePlatformConstraints(),
       logger: logger,
     );
   });
@@ -347,7 +349,6 @@ void main() {
         api: api,
         agUiClient: agUiClient,
         toolRegistry: _registryWith(),
-        platformConstraints: const NativePlatformConstraints(),
         logger: logger,
       );
       stubCreateRun();
@@ -378,7 +379,6 @@ void main() {
         api: api,
         agUiClient: agUiClient,
         toolRegistry: _registryWith(toolName: 'other_tool'),
-        platformConstraints: const NativePlatformConstraints(),
         logger: logger,
       );
       stubCreateRun();
@@ -420,7 +420,6 @@ void main() {
         api: api,
         agUiClient: agUiClient,
         toolRegistry: _registryWith(),
-        platformConstraints: const NativePlatformConstraints(),
         logger: logger,
       );
       stubCreateRun();
@@ -473,7 +472,6 @@ void main() {
         api: api,
         agUiClient: agUiClient,
         toolRegistry: _registryWith(),
-        platformConstraints: const NativePlatformConstraints(),
         logger: logger,
       );
       stubCreateRun();
@@ -516,7 +514,6 @@ void main() {
         api: api,
         agUiClient: agUiClient,
         toolRegistry: _registryWith(),
-        platformConstraints: const NativePlatformConstraints(),
         logger: logger,
       );
       stubCreateRun();
@@ -553,7 +550,6 @@ void main() {
         api: api,
         agUiClient: agUiClient,
         toolRegistry: _registryWith(),
-        platformConstraints: const NativePlatformConstraints(),
         logger: logger,
       );
       stubCreateRun();
@@ -575,7 +571,6 @@ void main() {
         api: api,
         agUiClient: agUiClient,
         toolRegistry: _registryWith(),
-        platformConstraints: const NativePlatformConstraints(),
         logger: logger,
       );
       stubCreateRun();
@@ -632,7 +627,6 @@ void main() {
         api: api,
         agUiClient: agUiClient,
         toolRegistry: _registryWith(),
-        platformConstraints: const NativePlatformConstraints(),
         logger: logger,
       );
       stubCreateRun();
