+++ packages/soliplex_client/test/application/run_state_test.dart
@@ -1,5 +1,7 @@
-import 'package:soliplex_agent/soliplex_agent.dart';
 import 'package:soliplex_client/soliplex_client.dart';
+import 'package:soliplex_client/src/application/run_state.dart';
+import 'package:soliplex_client/src/domain/failure_reason.dart';
+import 'package:soliplex_client/src/domain/thread_key.dart';
 import 'package:test/test.dart';
 
 const ThreadKey _key = (
