+++ packages/soliplex_client/test/domain/thread_key_test.dart
@@ -1,4 +1,4 @@
-import 'package:soliplex_agent/soliplex_agent.dart';
+import 'package:soliplex_client/src/domain/thread_key.dart';
 import 'package:test/test.dart';
 
 void main() {
